Auxiliary Functions and Utilities

Auxiliary Functions and Utilities

Linear Algebra

Loss Functions

Adding Bias

Other Utilities

Linear Algebra

It is essential to be able solve for the updated $\boldsymbol{\beta}$ when we update the $\boldsymbol{\omega}$. In a typical setting, this corresponds to solving

\[({S^\ast}S + N \lambda I)\boldsymbol{\beta} = S^\ast \boldsymbol{y},\]

We have included two naive solvers for this problem:

ARFF.solve_normal! — Function

solve_normal!(β, S, y_data; λ = 1e-8)

Solve the regularized linear system using the normal equations.

Fields

β - The vector of coefficients that will be obtained

S - The design matrix

y_data - y coordinates

λ = 1e-8 - Regularization parameter

ARFF.solve_normal_svd! — Function

solve_normal_svd!(β, S, y_data; λ = 1e-8)

Solve the regularized linear system using the SVD.

Fields

β - The vector of coefficients that will be obtained

S - The design matrix

y_data - y coordinates

λ = 1e-8 - Regularization parameter

Other formulations may be more appropriate. Indeed, in [2], the authors use the regularized loss function in the spirit of Sobolev:

\[\|S\boldsymbol{\beta}-\boldsymbol{y}\|_2^2 + \lambda \sum_{k} (1+|\omega_k|^2)|\beta_k|^2.\]

We thus require that the user provided linear solver take the form:

function linear_solver!(β, S, y, ω)
    # solve for β coefficients
    β
end

Obviously, if one does not need ω for your formulation, as is the case in the original regularization of [1], this argument is just ignored. For solve_normal! we would implement this as:

linear_solver! = (β, S, y, ω) -> solve_normal!(β, S, y)

In the case that one is solving a vector valued problem, the vector valued $\beta$'s are obtained component by component against the vector valued $y$'s.

Loss Functions

ARFF.mse_loss — Function

mse_loss(F::ScalarFourierModel{TC,TR,TW}, data_x::Vector{TW}, data_y::Vector{TC}) where {TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}}

Mean squared error loss function

Fields

F - A FourierModel structure

data_x - the x coordinates in training data

data_y - the y coordinates in training data

mse_loss(F::VectorFourierModel{TC,TR,TW,TB}, data_x::Vector{TW}, data_y::Vector{TB}) where {TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}, TB<:AbstractArray{TC}}

Mean squared error loss function

Fields

F - A FourierModel structure

data_x - the x coordinates in training data

data_y - the y coordinates in training data

As the entire framework is built around the mean square loss function, we have included it for convenience. Other loss functions can be implemented, but they should have the calling sequence:

function loss_function(F, data_x, data_y)
    # compute loss 
    return loss
end

Adding Bias

It may be neccessary to modify an existing data set so as to include a constant bias term in a model. If $x\in \mathbb{R}^d$, then

\[x\mapsto (x,1)=\tilde{x}\in \mathbb{R}^{d+1}.\]

This is relevant when using generalized activation functions; see Scalar Example with Generalized Activation Functions. We provide tools for account for the constant in both the DataSet and DataScalings types:

ARFF.append_bias — Method

append_bias(data)

Create a new DataSet with a bias term appended to the end of the x coordinate.

Fields

data - data set to be augmented

ARFF.append_bias — Method

append_bias(scalings)

Create a new DataScalings accounting for a bias term appended to the end of the x coordinate.

Fields

scalings - scalings to be augmented

In this framework, we would compute and/or apply the scalings before adding the bias term to the data set, and then add the bias in to both the data and the scaling structures. This avoids a potential divide by zero issue.

n_x = 10
x = [Float64[i] for i in 1:n_x]
f(x) = x[1]^2
y = f.(x)
data_ = DataSet(x, y)
scalings_ = get_scalings(data_);
data_scaled = scale_data!(data_, scalings_);

ARFF.ScalarDataSet{Float64, Float64, Int64}([[-1.4863010829205867], [-1.1560119533826787], [-0.8257228238447705], [-0.49543369430686224], [-0.1651445647689541], [0.1651445647689541], [0.49543369430686224], [0.8257228238447705], [1.1560119533826787], [1.4863010829205867]], [-1.0973390496411197, -1.00955192566983, -0.8632400523843474, -0.6584034297846718, -0.39504205787080304, -0.07315593664274131, 0.30725493389951347, 0.7461905537559613, 1.2436509229266022, 1.7996360414114363], 10, 1)

The means and variances of the scaled data set are as we would expect:

println(mean(data_scaled.x));
println(var(data_scaled.x));

[0.0]
[1.0]

and the scalings_ has the relevant information:

println(scalings_.μx);
println(scalings_.σ2x);

[5.5]
[9.166666666666666]

Next, we can add in the bias into the data set and check that it is computing properly:

data_bias = append_bias(data_scaled);
println(mean(data_bias.x));
println(var(data_bias.x));

[0.0, 1.0]
[1.0, 0.0]

and analogously with the scalings:

scalings_bias = append_bias(scalings_);
println(scalings_bias.μx);
println(scalings_bias.σ2x);

[5.5, 0.0]
[9.166666666666666, 1.0]

Other Utilities

ARFF.optimal_γ — Function

optimal_γ(d)

Compute the optimal γ parameter as a function of dimension d

Fields

d - the dimension of the x coordinate

Following Remark 1 in [1], the optimal $\gamma$ corresponds to $\gamma = 3d -2$, which is encoded in the above function.