Data Sets

Training (and testing data) are stored in a DataSet structure (with either scalar or vector valued $y$)

ARFF.ScalarDataSet — Type

ScalarDataSet{TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TI<:Integer}

Training data containing (x,y) data pairs stored in arrays of x values and arrays of y values.

Fields

x - Array of real valued vectors

y - Array of complex scalars

N - Number of data points

dx - Dimension of x coordinates

ARFF.VectorDataSet — Type

VectorDataSet{TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TB<:AbstractArray{TC},TI<:Integer}

Training data containing (x,y) data pairs stored in arrays of x values and arrays of y values.

Fields

x - Array of real valued vectors

y - Array of complex valued vectors

yt - Tranposed array of complex valued vectors

N - Number of data points

dx - Dimension of x coordinates

dy - Dimension of y coordinates

Constructing Data Sets

Data sets can be constructed by passing in the $(x_i,y_i)$ information:

ARFF.DataSet — Function

DataSet(x::Vector{Vector{TR}}, y::Vector{TY}) where {TR<:AbstractFloat,TY<:Number}

Convenience constructor for a scalar valued data set

Fields

x - Array of real valued vectors

y - Array of scalars

DataSet(x::Vector{Vector{TR}}, y::Vector{Vector{TY}}) where {TY<:Number, TR<:AbstractFloat}

Convenience constructor for a vector valued data set

Fields

x - Array of real valued vectors

y - Array of vectors

For instance, a scalar valued data set can be generated with: A training set can be constructed as follows:

using ARFF
using Random
Random.seed!(100); # for reproducibility

n = 10; # number of sample points
x = [rand(2) for _ in 1:n];
f(x) = exp(-x[1]*x[2]); # arbitrary function
y = complex.(f.(x)); # make y valued data complex for type consistency

data = DataSet(x,y);
println(data.x)
println(data.y)

[[0.12364754150469992, 0.17728426618671156], [0.5844262152209776, 0.9102808037366388], [0.04170637082547557, 0.2882051781587147], [0.8650936306221244, 0.8445261118175335], [0.5088587881265387, 0.021567602762348703], [0.7211452674280416, 0.5056618567185127], [0.31371781140887434, 0.34261122353631823], [0.2325290249107962, 0.5860467972325104], [0.8793427937731793, 0.17708124683920934], [0.6808445302529258, 0.3737483795813732]]
ComplexF64[0.9783177502966277 + 0.0im, 0.587433656224266 + 0.0im, 0.9880519594959999 + 0.0im, 0.48162274398150406 + 0.0im, 0.98908513990478 + 0.0im, 0.6944357352604994 + 0.0im, 0.8980915714548457 + 0.0im, 0.8726044746868709 + 0.0im, 0.8558029582309944 + 0.0im, 0.7753315459150839 + 0.0im]

Vector valued data can be similarly constructed:

using ARFF
using Random
Random.seed!(100); # for reproducibility

n = 10; # number of sample points
x = [rand(2) for _ in 1:n];
f(x) = [x[1]+x[2]; exp(-x[1]*x[2])]; # arbitrary function
y = f.(x);

data = DataSet(x,y);
println(data.x)
println(data.y)

[[0.12364754150469992, 0.17728426618671156], [0.5844262152209776, 0.9102808037366388], [0.04170637082547557, 0.2882051781587147], [0.8650936306221244, 0.8445261118175335], [0.5088587881265387, 0.021567602762348703], [0.7211452674280416, 0.5056618567185127], [0.31371781140887434, 0.34261122353631823], [0.2325290249107962, 0.5860467972325104], [0.8793427937731793, 0.17708124683920934], [0.6808445302529258, 0.3737483795813732]]
[[0.3009318076914115, 0.9783177502966277], [1.4947070189576164, 0.587433656224266], [0.3299115489841903, 0.9880519594959999], [1.7096197424396578, 0.48162274398150406], [0.5304263908888874, 0.98908513990478], [1.2268071241465544, 0.6944357352604994], [0.6563290349451926, 0.8980915714548457], [0.8185758221433066, 0.8726044746868709], [1.0564240406123886, 0.8558029582309944], [1.054592909834299, 0.7753315459150839]]

Scaling Data Sets

It is often helpful to scale the training data (setting means to zero and variances to unity). These can be accomplished with the scalings structure and the associated functions.

ARFF.get_scalings — Function

get_scalings(data::ScalarDataSet{TC,TR,TW,TI}) where {TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TI<:Integer}

Find the means and variances of the data for scaling

Fields

data - The training data set

get_scalings(data::VectorDataSet{TC,TR,TW,TB,TI}) where {TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TB<:AbstractArray{TC},TI<:Integer}

Find the means and variances of the data for scaling

Fields

data - The training data set

ARFF.scale_data! — Function

scale_data!(data::ScalarDataSet{TC,TR,TW,TI}, scalings::ScalarDataScalings{TC,TR,TW}) where {TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TI<:Integer}

Scale the data set (in-place) according to the specified scalings

Fields

data - Data set to be scale

scalings - Scalings to apply to data

scale_data!(data::VectorDataSet{TC,TR,TW,TB,TI}, scalings::VectorDataScalings{TC,TR,TW,TB}) where {TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}, TB<:AbstractArray{TC},TI<:Integer}

Scale the data set (in-place) according to the specified scalings

Fields

data - Data set to be scale

scalings - Scalings to apply to data

ARFF.rescale_data! — Function

rescale_data!(data::ScalarDataSet{TC,TR,TW}, scalings::ScalarDataScalings{TC,TR,TW}) where {TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}}

Rescale the data set (in-place) according back to the original units

Fields

data - Data set to be scale

scalings - Scalings to apply to data

rescale_data!(data::VectorDataSet{TC,TR,TW,TB}, scalings::VectorDataScalings{TC,TR,TW,TB}) where {TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}, TB<:AbstractArray{TC}}

Rescale the data set (in-place) according back to the original units

Fields

data - Data set to be scale

scalings - Scalings to apply to data

For example, the following code obtains the means and variances in the data:

scalings = get_scalings(data);
println(scalings.μx);
println(scalings.σ2x);
println(scalings.μy);
println(scalings.σ2y);

[0.4951311974073633, 0.4227013466569871]
[0.09152105546137086, 0.08417790437576789]
0.8120777535451472 + 0.0im
0.030862604324702423

Next, we can scale our data:

scale_data!(data, scalings);

using Statistics
println(mean(data.x));
println(var(data.x));
println(mean(data.y));
println(var(data.y));

[2.220446049250313e-16, 5.551115123125783e-18]
[1.0000000000000002, 1.0]
-1.0824674490095276e-16 + 0.0im
1.0

If needed, we can then undo the scaling,

rescale_data!(data, scalings)

println(mean(data.x));
println(var(data.x));
println(mean(data.y));
println(var(data.y));

[0.4951311974073633, 0.4227013466569871]
[0.09152105546137086, 0.08417790437576789]
0.8120777535451472 + 0.0im
0.030862604324702423

Note that these are in agreement with what was contained in our scalings structure.