Fourier Feature Models

The generalized fourier features model approximates functions with $K$ features, with acvtivation function $\varphi$ as

\[f_{\rm true}(x) \approx f(x) = \sum_{k=1}^K \beta_k \varphi(x;\omega_k).\]

In the above expression, $x,\omega_k\in \mathbb{R}^d$, while $\beta_k\in\mathbb{R}^{d'}$ or $\beta_k\in\mathbb{C}^{d'}$, depending on the choice of activation function. Consequently, our model is uniquely determined by the coefficients and the wavenumbers.

These are stored in either a scalar or vector valued data structure:

ARFF.ScalarFourierModel — Type

ScalarFourierModel{TB<:Complex,TW<:AbstractArray{AbstractFloat}}

Structure containing a scalar valued fourier model which will be learned

Fields

β - Array of complex coefficients

ω - Array of wave vectors

K - Number of Fourier features

dx - Dimension of x coordinate

ϕ - Activation function

ARFF.VectorFourierModel — Type

VectorFourierModel{TB<:Complex,TW<:AbstractArray{AbstractFloat}}

Structure containing a scalar valued fourier model which will be learned

Fields

β - Array of complex coefficients

βt - Transposed array of complex coefficients

ω - Array of wave vectors

K - Number of Fourier features

dx - Dimension of x coordinate

dy - Dimension of y coordinate

Constructing a Fourier Features Model

A Fourier features model can be instantiated with:

ARFF.FourierModel — Function

FourierModel(β::Vector{TR}, ω::Vector{Vector{TR}}) where {TR<:AbstractFloat}

TBW

FourierModel(β::Vector{TB}, ω::Vector{Vector{TR}}, ϕ::TA) where {TB<:Number,TR<:AbstractFloat,TA<:ActivationFunction{TB}}

TBW

FourierModel(β::Vector{Vector{TR}}, ω::Vector{Vector{TR}}) where {TR<:AbstractFloat}

TBW

FourierModel(β::Vector{Vector{TB}}, ω::Vector{Vector{TR}}, ϕ::TA) where {TR<:AbstractFloat,TB<:Number,TA<:ActivationFunction{TB}}

TBW

As an example, the scalar valued Fourier model with complex exponentials can be instantiated as:

using ARFF
using Random
Random.seed!(200); # for reproducibility

K = 10;
d = 2;
# convenience
F = FourierModel(randn(ComplexF64, K),  [randn(d) for _ in 1:K]);

ARFF.ScalarFourierModel{Float64, ComplexF64, Int64, ActivationFunction{ComplexF64, typeof(ARFF.fourier)}}(ComplexF64[-0.5603272581116238 + 0.537977815860294im, -0.5706975858116924 - 0.0657348379263049im, 0.9069388509515408 + 0.9326057155997285im, 1.515643502824237 - 0.628782197570827im, 0.6562851968100064 - 0.2961858090925524im, -0.0826739515401274 - 0.022802052525077828im, 0.604928870663104 - 0.27642648514635815im, -1.0464139434115503 + 0.7296373349276786im, -0.48881518199590523 - 0.8415598542644231im, 1.7423067046684735 - 0.46776727576938804im], [[0.8446781051997678, 0.9249237650541624], [-0.5841871308675767, 0.3424865252782795], [-0.6974350383322714, 1.1297139822413678], [-1.243194987042106, -0.9922723588323762], [1.9537974625817172, -0.18686037485344426], [-1.146293771939666, 0.9226286258452874], [0.2055244369510107, 0.13819789976281208], [-0.7545830090756918, -0.8047536925085138], [-0.40084633859373503, 0.9736129994729481], [1.1241570456776535, 0.012007461271785093]], 10, 2, ActivationFunction{ComplexF64, typeof(ARFF.fourier)}(ARFF.fourier))

This defines F with random wavenumbers and amplitudes. Strictly speaking, this is not required, but it is often helpful within the learning context that we will apply this method. By not specifying an activation function, this defaults to FourierActivation function, e^{i \omega \cdot x}, and presumes all y related values are of the same complex type.

Function evaluation has been overloaded for a FourierModel, allowing us to evaluate it at points:

x_test = [0., 1.];
println(F(x_test));

1.7282986547022365 - 1.519135174016236im

and it can perform a vectorized evaluation:

n = 10; # number of sample points
x = [rand(2) for _ in 1:n];
f(x) = exp(-x[1]*x[2]); # arbitrary function
y = complex.(f.(x)); # make y valued data complex for type consistency

data = DataSet(x,y);

F.(data.x);

10-element Vector{ComplexF64}:
 2.0916342737755884 + 0.11015373831256492im
 0.7407822793578944 + 1.3446409542541862im
  2.309270885469973 + 0.17868415429852003im
 2.1001930318974136 - 0.9570423428141069im
  1.355761770475617 + 1.014245254983245im
 1.8081509272040823 - 0.3036415530293624im
 1.8794300217862092 - 1.0306918361668578im
 1.8208390394443557 - 0.0864377833507366im
  2.484791666443803 + 0.37182849619748487im
 0.9036171113723462 + 1.3222505463363432im

As is sometimes the case, we may scale our data, train in the scale coordinate, and wish to evaluate new poitns in the original, unscaled, coordinate system. We can accomplish that by passing a DataScalings argument in for evaluation:

scalings = get_scalings(data);

F(x_test, scalings)

0.7706115707461505 - 0.1778212908233697im

If one wishes to use a different activation function, we pass that as an argument in the construction:

G = FourierModel(randn(K),  [randn(d) for _ in 1:K],SigmoidActivation);
println(G([0.1, 0.2]))

1.7083831207492426