Training

Training

In Place Training

Recording the Training Trajectory

The key function is train_rwm!, which performs in place training on the model. This is implemented to handle the training data in several ways:

A single DataSet can be provided and used in every training epoch.

A single DataSet and a minibatch size can be provided, and minibatchs will be generated at each epoch.

An array of DataSet types can be provided, and they will be cycled through each epoch;

In Place Training

For both the scalar and vector valued case, we have the following commands:

ARFF.train_rwm! — Method

train_rwm!(F, data, Σ, options; show_progress=true, record_loss=true)

F - The FourierModel to be trained

data- A DataSet training data set

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_rwm! — Method

train_rwm!(F, data, batch_size, Σ, options; show_progress=true, record_loss=true)

F - The FourierModel to be trained

data- A DataSet training data set

batch_size - Minibatch size

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_rwm! — Method

train_rwm!(F, data_sets, Σ, options; show_progress=true, record_loss=true)

F - The FourierModel to be trained

data_sets- A vector of DataSet training data sets. These are presumed to all be the same size. They will be cycled through each epoch.

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

Having created an initial F we can then call

Σ_mean, acceptance_rate, loss = train_rwm!(F, data, Σ0, opts);

The returned quantities are the mean adapted covariance matrix Σ_mean. The acceptance_rate is the mean acceptance rate at each epoch, averaged overa the internal steps, K * n_ω_steps. The loss is the recorded training loss, stored in opts.loss, at each epoch.

Recording the Training Trajectory

We have also included routines which record the model at each epoch. These are called in the same way as above:

F_trajectory, Σ_mean, acceptance_rate, loss = train_rwm(F0, data, Σ0, opts);

The functions are analogously named:

ARFF.train_rwm — Method

train_rwm(F₀, data, Σ, options; show_progress=true, record_loss=true)

F₀ - The initial state of the FourierModel to be trained

data- A DataSet training data set

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_rwm — Method

train_rwm(F₀, data, batch_size, Σ, options; show_progress=true, record_loss=true)

F₀ - The initial state of the FourierModel to be trained

data- A DataSet training data set

batch_size - Minibatch size

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_rwm — Method

train_rwm(F₀, data_sets, Σ, options; show_progress=true, record_loss=true)

F₀ - The initial state of the FourierModel to be trained

data_sets- A vector of DataSet training data sets. These are presumed to all be the same size. They will be cycled through each epoch.

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

This records the result at the end of each epoch.