General Training

Data Structure

In the most general case, when a user calls train_arff! or train_arff, they must provide an ARFFSolver data structure:

ARFF.ARFFSolver — Type

ARFFSolver{TS, TM, TR, TL}

Data structure containing key parameters for ARFF training

Fields

linear_solve! - User specified solver for the normal equations

mutate! - In place transformation for the mutation/exploration step

resample! - In place transformation for the resampling step

loss - Loss function

This includes stores the following functions:

linear_solve!(β, ω, x, y, S, epoch): β and ω are the coefficients and wave numbers; (x,y) are arrays of training pairs; S is the design matrix; and epoch is the current epoch. Not all arguments are neccessarily used, but this is how the solver will be called.

mutate!(F, x, y, S, epoch): F is the Fourier model; (x,y) are arrays of training pairs; S is the design matrix; and epoch is the current epoch.

resample!(F, x, y, S, epoch): This follows the convetion of mutate!

loss(F, x, y): F is the Fourier model; and (x,y) are arrays of training pairs.

Training

These are underlying, general, training functions. The user may wish to access these if they have custom mutation and resampling steps. Otherwise, see, for instance Random Walk Metropolis Training which performs the algorithm described in [1]

ARFF.train_arff! — Method

train_arff!(F, data_sets, batch_size, solver, n_epochs; show_progress=true, record_loss=true)

Perform RWM training in place on an ARFF model

Fields

F - Fourier feature model to be trained, in place

data_sets - Training data sets

batch_size - Minibatch size

solver - An ARFFSolver data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_arff — Method

train_arff(F₀, data_sets, batch_size, solver, n_epochs; show_progress=true, record_loss=true)

Perform RWM training an ARFF model, recording the result at all epochs.

Fields

F₀ - Initial fourier feature model to be trained

data_sets - Training data sets

batch_size - Minibatch size

solver - An ARFFSolver data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

For convenience, we have also included the following versions of these functions:

ARFF.train_arff! — Method

train_arff!(F, data, solver, n_epochs; show_progress=true, record_loss=true)

Perform ARFF training on a given data set

Fields

F - Fourier feature model to be trained, in place

data - Training data set

solver - An ARFFSolver data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_arff! — Method

train_arff!(F, data, solver, n_epochs; show_progress=true, record_loss=true)

Perform ARFF training on a given data set with minibatching

Fields

F - Fourier feature model to be trained, in place

data - Training data set

batch_size - Minibatch size

solver - An ARFFSolver data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_arff! — Method

train_arff!(F, data, solver, n_epochs; show_progress=true, record_loss=true)

Perform ARFF training on a given data set with minibatching

Fields

F - Fourier feature model to be trained, in place

data_sets - Array of training data sets

solver - An ARFFSolver data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_arff — Method

train_arff(F₀, data, solver, n_epochs; show_progress=true, record_loss=true)

Perform RWM training an ARFF model, recording the result at all epochs.

Fields

F₀ - Initial fourier feature model to be trained

data - Training data set

solver - An ARFFSolver data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_arff — Method

train_arff(F₀, data, batch_size, solver, n_epochs; show_progress=true, record_loss=true)

Perform RWM training an ARFF model, recording the result at all epochs.

Fields

F₀ - Initial fourier feature model to be trained

data - Training data set

batch_size - Minibatch size

solver - An ARFFSolver data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_arff — Method

train_arff(F₀, data_sets, solver, n_epochs; show_progress=true, record_loss=true)

Perform RWM training an ARFF model, recording the result at all epochs.

Fields

F₀ - Initial fourier feature model to be trained

data_sets - Array of training data sets

solver - An ARFFSolver data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record