Random Walk Metropolis Training

For convenience, we have implemented the RWM strategy from [1].

RWM Data Structures

Classical RWM

For a classical RWM, we can employ a RWMSampler data structure, which can be constructed with one of the following commands:

ARFF.RWMSampler — Method

RWMSampler(F, linear_solve!, n_rwm_steps, δ)

Constructor for the RWMSampler data structure. Defaults to ω_max = Inf, γ = optimal_γ(dx), and Σ= I

Fields

F - Fourier feature model; used for setting types and dimensions

linear_solve! - User specified solver for the normal equations

n_rwm_steps - Number of internal RWM steps

δ - RWM proposal step size

ARFF.RWMSampler — Method

RWMSampler(F, linear_solve!, n_rwm_steps, Σ, γ, δ, ω_max)

Constructor for the RWMSampler data structure

Fields

F - Fourier feature model; used for setting types and dimensions

linear_solve! - User specified solver for the normal equations

n_rwm_steps - Number of internal RWM steps

Σ - Covariance matrix

γ - Metropolis-Hastings exponent

δ - RWM proposal step size

ω_max - Maximum wave number norm cutoff

The linear_solve! argument should be of the same type as in ARFFSolver, `linear_solve!(β, ω, x, y, S, epoch); see Linear Algebra for additional details. It is not neccessary to separately construct an ARFFSolver structure if using these. These default to the mean squared error loss function, mse_loss. As an exmaple, the following code is sufficient to construct such a structure, assuming one has already constructed an initial FourierModel, F0:

δ = 1.0 # rwm step size
n_rwm_steps = 10 # number of RWM steps within the mutation step

# format the linear solver function to be compatible with ARFF.jl
linear_solver! = (β, ω, x, y, S, epoch) -> solve_normal!(β, S, y)

rwm_sampler = RWMSampler(F0, linear_solver!, n_rwm_steps, δ);

RWMSampler{Main.var"#5#6", Int64, Matrix{Float64}, Float64, Distributions.ZeroMeanFullNormal{Tuple{Base.OneTo{Int64}}}, Vector{Vector{Float64}}, Vector{ComplexF64}}(Main.var"#5#6"(), 10, ComplexF64[-1.334432356273704 + 0.0im, -1.604490016105518 + 0.0im, 1.4102378010315544 + 0.0im, 0.8957459720530755 + 0.0im, 0.5932655467240973 + 0.0im, -0.8805550396393875 + 0.0im, 0.5110704161210543 + 0.0im, 1.426726767257056 + 0.0im, 0.7146501248388691 + 0.0im, 0.28719872597010615 + 0.0im  …  1.5373777895586518 + 0.0im, -0.6022001356422086 + 0.0im, 0.0002862652282358044 + 0.0im, -1.2199904108784254 + 0.0im, -0.10324186835603144 + 0.0im, 3.192135338124483 + 0.0im, 0.13907316355757052 + 0.0im, -0.06728909981123397 + 0.0im, -0.2115049356214002 + 0.0im, -0.9685487561234694 + 0.0im], [[-1.1070232525585124], [-0.7439202680807437], [0.20640021391145197], [-0.1404339880416136], [2.3966039097485723], [0.2741992933558498], [-0.9050125314083654], [-1.839301196577807], [0.9931072306471086], [1.082445975136616]  …  [-0.9968330013297498], [-0.37248908478292936], [-0.027124229247490077], [0.2951385291544223], [0.024676514448530786], [0.4108351363170801], [1.494635036056227], [0.48206961996619596], [1.3774697929360409], [-0.9000106604672613]], [1.0;;], 1, 1.0, Float64[], Inf, ZeroMeanFullNormal(
dim: 1
μ: Zeros(1)
Σ: [1.0;;]
)
)

Adaptive RWM

It may be more useful to use an AdaptiveRWMSampler data structure; this will update proposal covariance matrix after n_burn epochs. This is the version used in our examples.

The AdaptiveRWMSampler structure constructed with:

ARFF.AdaptiveRWMSampler — Method

AdaptiveRWMSampler(F, linear_solve!, n_rwm_steps, n_burn, δ)

Constructor for the AdaptiveRWMSampler data structure. Defaults to ω_max = Inf, γ = optimal_γ(dx), and Σ0= I

Fields

F - Fourier feature model; used for setting types and dimensions

linear_solve! - User specified solver for the normal equations

n_rwm_steps - Number of internal RWM steps

n_burn - Number of epochs before the covariance adaptation begins

δ - RWM proposal step size

ARFF.AdaptiveRWMSampler — Method

AdaptiveRWMSampler(F, linear_solve!, n_rwm_steps, n_burn, Σ0, γ, δ, ω_max)

Constructor for the AdaptiveRWMSampler data structure

Fields

F - Fourier feature model; used for setting types and dimensions

linear_solve! - User specified solver for the normal equations

n_rwm_steps - Number of internal RWM steps

n_burn - Number of epochs before the covariance adaptation begins

Σ0 - Initial covariance matrix

γ - Metropolis-Hastings exponent

δ - RWM proposal step size

ω_max - Maximum wave number norm cutoff

As an example:

δ = 1.0 # rwm step size
n_rwm_steps = 10 # number of RWM steps within the mutation step
n_burn = 100 # number of training epochs to use for burn in.

# format the linear solver function to be compatible with ARFF.jl
linear_solver! = (β, ω, x, y, S, epoch) -> solve_normal!(β, S, y)

rwm_sampler = AdaptiveRWMSampler(F0, linear_solver!, n_rwm_steps, n_burn, δ); nothing

RWM Mutation

As noted in General Training, the underlying training algorithm for ARFF requires the specification of a mutation step. When we use RWM for that mutation step, as is automatically done for the user in RWM Training, and many of our examples, what we are actually doing is first constructing the mutation function, and then embedding it within an ARFFSolver data structure, as follows:

mutate_rwm!(F, x, y, S, epoch) = ARFF.rwm!(F0, rwm_sampler, x, y, S, epoch);
solver = ARFFSolver(rwm_sampler.linear_solve!, mutate_rwm!, trivial_resample!, mse_loss); nothing

This example assumes, of course, that you have already constructed the rwm_sample object.

RWM Training

RWM training is implemented to handle the training data in several ways:

A single DataSet can be provided and used in every training epoch.

A single DataSet and a minibatch size can be provided, and minibatchs will be generated at each epoch.

An array of DataSet types can be provided, and they will be cycled through each epoch.

Additionally, we have both an in place training commands, along with commands which will record the training trajectory. These can be used for both scalar and vector valued problems.

ARFF.train_rwm! — Method

train_rwm!(F, data, rwm_sampler; show_progress=true, record_loss=true)

Perform RWM training of an ARFF model

Fields

F - Fourier feature model to be trained, in place

data - Training data

rwm_sampler - An RWM sampler data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_rwm! — Method

train_rwm!(F, data, batch_size, rwm_sampler, n_epochs; show_progress=true, record_loss=true)

Perform RWM training of an ARFF model with mini batching

Fields

F - Fourier feature model to be trained, in place

data - Training data

batch_size - Minibatch size

rwm_sampler - An RWM sampler data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_rwm! — Method

train_rwm!(F, data, batch_size, rwm_sampler, n_epochs; show_progress=true, record_loss=true)

Perform RWM training of an ARFF model with an array of different training data sets

Fields

F - Fourier feature model to be trained, in place

data_sets - Training data sets

rwm_sampler - An RWM sampler data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

Having created an initial F we can then call

acceptance_rate, loss = train_rwm!(F, data, rwm_sampler, n_epochs);

The acceptance_rate is the mean acceptance rate at each epoch, averaged over the internal steps, K * n_rwm_steps. The loss is the recorded training loss at each epoch. If using AdaptiveRWMSampler, the tuned covariance can be obtained by examining

rwm_sampler.Σ_mean

ARFF.train_rwm — Method

train_rwm(F₀, data, rwm_sampler; show_progress=true, record_loss=true)

Perform RWM training of an ARFF model

Fields

F₀ - Initial Fourier feature model to be trained

data - Training data

rwm_sampler - An RWM sampler data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_rwm — Method

train_rwm(F₀, data, batch_size, rwm_sampler, n_epochs; show_progress=true, record_loss=true)

Perform RWM training of an ARFF model with mini batching

Fields

F₀ - Initial Fourier feature model to be trained

data - Training data

batch_size - Minibatch size

rwm_sampler - An RWM sampler data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

ARFF.train_rwm — Method

train_rwm(F₀, data, batch_size, rwm_sampler, n_epochs; show_progress=true, record_loss=true)

Perform RWM training of an ARFF model with an array of different training data sets

Fields

F₀ - Initial Fourier feature model to be trained

data_sets - Training data sets

rwm_sampler - An RWM sampler data structure

n_epochs - Number of training epochs

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

Having created an initial F₀ we can then call

F_trajecotry, acceptance_rate, loss = train_rwm(FF₀, data, rwm_sampler, n_epochs);

One of the returned quantities is the algorithmic trajectory, F_trajecotry. The acceptance_rate is the mean acceptance rate at each epoch, averaged overa the internal steps, K * n_rwm_steps. The loss is the recorded training loss at each epoch.