Auxiliary Functions and Utilities

Auxiliary Functions and Utilities

Linear Algebra

Loss Functions

Other Utilities

Linear Algebra

It is essential to be able solve for the updated $\boldsymbol{\beta}$ when we update the $\boldsymbol{\omega}$. In a typical setting, this corresponds to solving

\[({S^\ast}S + N \lambda I)\boldsymbol{\beta} = S^\ast \boldsymbol{y},\]

We have included two naive solvers for this problem:

ARFF.solve_normal! — Function

solve_normal!(β, S, y_data; λ = 1e-8)

Solve the regularized linear system using the normal equations.

Fields

β - The vector of coefficients that will be obtained

S - The design matrix

y_data - y coordinates

λ = 1e-8 - Regularization parameter

ARFF.solve_normal_svd! — Function

solve_normal_svd!(β, S, y_data; λ = 1e-8)

Solve the regularized linear system using the SVD.

Fields

β - The vector of coefficients that will be obtained

S - The design matrix

y_data - y coordinates

λ = 1e-8 - Regularization parameter

Other formulations may be more appropriate. Indeed, in [2], the authors use the regularized loss function in the spirit of Sobolev:

\[\|S\boldsymbol{\beta}-\boldsymbol{y}\|_2^2 + \lambda \sum_{k} (1+|\omega_k|^2)|\beta_k|^2.\]

We thus require that the user provided linear solver take the form:

function linear_solver!(β, S, y, ω)
    # solve for β coefficients
    β
end

Obviously, if one does not need ω for your formulation, as is the case in the original regularization of [1], this argument is just ignored. For solve_normal! we would implement this as:

linear_solver! = (β, S, y, ω) -> solve_normal!(β, S, y)

In the case that one is solving a vector valued problem, the vector valued $\beta$'s are obtained component by component against the vector valued $y$'s.

Loss Functions

ARFF.mse_loss — Function

mse_loss(F::ScalarFourierModel{TC,TR,TW}, data_x::Vector{TW}, data_y::Vector{TC}) where {TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}}

Mean squared error loss function

Fields

F - A FourierModel structure

data_x - the x coordinates in training data

data_y - the y coordinates in training data

mse_loss(F::VectorFourierModel{TC,TR,TW,TB}, data_x::Vector{TW}, data_y::Vector{TB}) where {TC<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}, TB<:AbstractArray{TC}}

Mean squared error loss function

Fields

F - A FourierModel structure

data_x - the x coordinates in training data

data_y - the y coordinates in training data

As the entire framework is built around the mean square loss function, we have included it for convenience. Other loss functions can be implemented, but they should have the calling sequence:

function loss_function(F, data_x, data_y)
    # compute loss 
    return loss
end

Other Utilities

ARFF.optimal_γ — Function

optimal_γ(d::Integer)

Compute the optimal γ parameter as a function of dimension d

Fields

d - the dimension of the x coordinate

Following Remark 1 in [1], the optimal $\gamma$ corresponds to $\gamma = 3d -2$, which is encoded in the above function.