Data Structures

Data Structures

Data Sets

Fourier Feature Models

Training Options

Data Sets

Training (and testing data) are stored in DataSet structure.

ARFF.DataSet — Type

DataSet{TB<:Complex,TW<:AbstractArray{AbstractFloat}}

Training data containing (x,y) data pairs stored in arrays of x values and arrays of y values.

Fields

x - Array of real valued vectors

y - Array of complex scalars

A training set can be constructed as follows

using ARFF
using Random
Random.seed!(100); # for reproducibility

n = 10; # number of sample points
x = [rand(2) for _ in 1:n];
f(x) = exp(-x[1]*x[2]); # arbitrary function
y = f.(x);

data = DataSet(x,y);
println(data.x)
println(data.y)

[[0.2601250914736861, 0.19031281518127185], [0.6609109399808133, 0.06719317094984745], [0.9675998379215747, 0.6456910432314067], [0.5459681993995775, 0.5268448425879988], [0.9727545053153894, 0.8681941503861808], [0.17270743062211147, 0.5239900463071787], [0.5578372055383092, 0.8931692669285218], [0.19938059113361262, 0.08937190378164739], [0.3073738264292152, 0.32430299502917914], [0.6885052016067101, 0.8697162407893]]
ComplexF64[0.9517002678523827 + 0.0im, 0.9565629286067209 + 0.0im, 0.5353842589990513 + 0.0im, 0.7500311574329656 + 0.0im, 0.4297550983446796 + 0.0im, 0.9134770975584506 + 0.0im, 0.6075972416993739 + 0.0im, 0.9823387969940219 + 0.0im, 0.9051249735419101 + 0.0im, 0.5494683219266783 + 0.0im]

Additionally, it is often helpful to scale the training data (setting means to zero and variances to unity). These can be accomplished with the DataScalings structure and the associated functions.

ARFF.DataScalings — Type

DataScalings{TB<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}}

Data structure holding the scalings of a DataSet type. Used for centering and scaling the data to improve training.

Fields

μx - Mean in x

σ2x - Variance in x

μy - Mean in y

σ2y - Variance in y

ARFF.get_scalings — Function

get_scalings(data::DataSet{TB,TR,TW}) where {TB<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}}

Find the means and variances of the data for scaling

Fields

data - The training data set

ARFF.scale_data! — Function

scale_data!(data::DataSet{TB,TR,TW}, scalings::DataScalings{TB,TR,TW}) where {TB<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}}

Scale the data set (in-place) according to the specified scalings

Fields

data - Data set to be scale

scalings - Scalings to apply to data

ARFF.rescale_data! — Function

rescale_data!(data::DataSet{TB,TR,TW}, scalings::DataScalings{TB,TR,TW}) where {TB<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR}}

Rescale the data set (in-place) according back to the original units

Fields

data - Data set to be scale

scalings - Scalings to apply to data

For example, the following code obtains the means and variances in the data:

scalings = get_scalings(data);
println(scalings.μx);
println(scalings.σ2x);
println(scalings.μy);
println(scalings.σ2y);

[0.5333162829420999, 0.49987864751725325]
[0.08778541249005575, 0.1035637532134563]
0.7581440142956235 + 0.0im
0.04406289344932272

Next, we can scale our data:

scale_data!(data, scalings);

using Statistics
println(mean(data.x));
println(var(data.x));
println(mean(data.y));
println(var(data.y));

[2.7755575615628914e-16, 6.661338147750939e-17]
[1.0000000000000002, 1.0]
-4.4408920985006264e-17 + 0.0im
1.0

If needed, we can then undo the scaling,

rescale_data!(data, scalings)

println(mean(data.x));
println(var(data.x));
println(mean(data.y));
println(var(data.y));

[0.5333162829420999, 0.49987864751725325]
[0.08778541249005575, 0.10356375321345629]
0.7581440142956235 + 0.0im
0.04406289344932272

Note that these are in agreement with what was contained in our scalings structure.

Fourier Feature Models

Fourier features approximates functions with $K$ features according to

\[f_{\rm true}(x) \approx f(x) = \sum_{k=1}^K \beta_k e^{i \omega_k \cdot x}\]

In the above expression, $x,\omega_k\in \mathbb{R}^d$, while $\beta_k\in\mathbb{C}$. Consequently, our model is uniquely determined by the coefficients and the wavenumbers.

A Fourier features model can be instantiated with a FourierModel structure:

ARFF.FourierModel — Type

FourierModel{TB<:Complex,TW<:AbstractArray{AbstractFloat}}

Structure containing a scalar valued fourier model which will be learned

Fields

β - Array of complex coefficients

ω - Array of wave vectors

As an example,

Random.seed!(200); # for reproducibility

K = 10;
d = 2;

F = FourierModel([randn(ComplexF64) for _ in 1:K],
    [randn(d) for _ in 1:K]);

FourierModel{ComplexF64, Float64, Vector{Float64}}(ComplexF64[0.8538588234195934 + 0.09233396678855635im, 0.6657725001419279 + 0.13206162020440187im, 0.3729829595288897 + 0.8853733583750293im, -0.023607110384581298 - 0.3074390400673215im, 0.7392385795213402 - 0.7716314720318435im, 0.41411683967995017 + 0.42659055982171135im, -1.0899699679707044 - 1.2870743462795848im, -0.3338915487914278 - 0.4646422764616553im, 0.32659554982937244 - 0.1589836264316281im, 0.556467377796313 - 0.08223415989499312im], [[0.04626904992355818, 0.39535980010424693], [-1.5950793710251607, -0.14318009205294555], [0.5594285840722241, 0.03459095158608406], [-0.2422615110015699, 0.16233267011838423], [-1.309584356610609, -0.2528260502836213], [0.5886366036339431, -0.7308495188756913], [-0.30044543447177563, -0.20888361570685232], [-1.563270931769172, 1.082218403400999], [1.4781012065476677, -0.7150656318423128], [1.0269037689834073, -0.8546402111927031]])

This defines F with random wavenumbers and amplitudes. Strictly speaking, this is not required, but it is often helpful within the learning context that we will apply this method. Function evaluation has been overloaded for a FourierModel, allowing us to evaluate it at points:

x_test = [0., 1.];
println(F(x_test));

2.280614775214353 - 2.2044046447847947im

and it can perform a vectorized evaluation:

F.(data.x);

10-element Vector{ComplexF64}:
  1.7655934517128418 - 1.572112938608091im
   0.278367670021155 - 0.8353046386069838im
 -0.8709301310527467 - 0.5174698284233284im
  0.7674827850127541 - 1.5143382942678385im
 -0.8686087510294098 - 0.683800170565174im
   2.035343306106639 - 1.9671138298425686im
  0.7023336301941535 - 1.8256972042495492im
   1.945583295839178 - 1.5161294829742478im
   1.619587266164578 - 1.6653742093589807im
 0.21684888748622722 - 1.5247878383892501im

As is sometimes the case, we may scale our data, train in the scale coordinate, and wish to evaluate new poitns in the original, unscaled, coordinate system. We can accomplish that by passing a DataScalings argument in for evaluation:

F(x_test, scalings)

0.7150068150527544 + 0.08771547401096434im

Training Options

To train a FourierModel, the training options must first be specified by instantiating an ARFFOptions struct with the desired parameters.

ARFF.ARFFOptions — Type

ARFFOptions{TI<:Integer,TF<:AbstractFloat,TB<:Bool,TL,TS}

Data structure containing the training options and parameters

Fields

n_epochs - Number of training epochs

n_ω_steps - Number of internal RWM steps

δ - RWM proposal step size

n_burn - Number of epochs before the covariance adaptation begins

γ - Metropolis-Hastings exponent

ω_max - Maximum wave number norm cutoff

adapt_covariance - Boolean for adaptivity

linear_solve! - User specified solver for the normal equations

loss - User specified loss function

The above arguments are passed in as follows:

δ = 1.; # rwm step size
n_epochs = 10^3;
n_ω_steps = 10;
n_burn = 100;
γ = 1;
ω_max =Inf;
adapt_covariance = true;

# solve the  normal equations
function reg_β_solver!(β, S, y, λ)
    N = length(y);
    β .= (S' * S + λ * N *I) \ (S' * y);
    β
end

linear_solve! = (β, S, y, ω)-> reg_β_solver!(β, S, y, λ);

opts = ARFFOptions(n_epochs, n_ω_steps, δ, n_burn, γ, ω_max, adapt_covariance,
    linear_solve!, ARFF.mse_loss);

ARFFOptions{Int64, Float64, Bool, typeof(ARFF.mse_loss), Main.var"#7#8"}(1000, 10, 1.0, 100, 1, Inf, true, Main.var"#7#8"(), ARFF.mse_loss)

The linear solver type is typed as:

function linear_solve!(β, S, y, ω)
    # in place solver code...
end

which is to say, it can depend on the design matrix $S$, the response variables, $y$, and, potentially, the current set of wave numbers, $ω$. The loss function is to be typed as:

function loss(F, x, y)
    # compute and return loss function...
end

The γ parameter can be set using optimal_γ, which, in a certain regime, is the optimal value for given dimension, d when $x\in\mathbb{R}^d$.