Training

The key function is train_rwm!, which performs in place training on the model. This is implemented to handle the training data in several ways:

A single DataSet can be provided and used in every training epoch.

An array of DataSet types can be provided, and they will be cycled through each epoch;

A single DataSet and a minibatch size can be provided, and minibatchs will be generated at each epoch.

In Place Training

ARFF.train_rwm! — Function

train_rwm!(F::FourierModel{TB,TR,TW}, data::DataSet{TB,TR,TW}, Σ::TM, options::ARFFOptions; show_progress=true, record_loss=true) where {TB<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TM<:AbstractMatrix}

Train the Fourier feature model using a random walk Metropolis exploration strategy

Fields

F - The FourierModel to be trained

data- The DataSet training data

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

train_rwm!(F::FourierModel{TB,TR,TW}, batched_data::Vector{DataSet{TB,TR,TW}}, Σ::TM, options::ARFFOptions) where {TB<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TM<:AbstractMatrix}

Train the Fourier feature model using a random walk Metropolis exploration strategy with batched data, which is cycled through from epoch to epoch.

Fields

F - The FourierModel to be trained

batched_data- A vector of DataSet training data sets, for the purpose of minibatching. These are presumed to all be the same size.

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

train_rwm!(F::FourierModel{TB,TR,TW}, data::DataSet{TB,TR,TW}, batch_size::TI, Σ::TM, options::ARFFOptions; show_progress=true, record_loss=true) where {TB<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TM<:AbstractMatrix,TI<:Integer}

Train the Fourier feature model using a random walk Metropolis exploration strategy with minibatching, randomly subsampling at each epoch.

F - The FourierModel to be trained

batch_size- Minibatch size

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

Having created an initialwe can then call

Σ_mean, acceptance_rate, loss = train_rwm!(F, data, Σ0, opts);

The returned quantities are the mean adapted covariance matrix Σ_mean. The acceptance_rate is the mean acceptance rate at each epoch, averaged overa the internal steps, K * n_ω_steps. The loss is the recorded training loss, stored in opts.loss, at each epoch.

Recording the Training Trajectory

We have also included routines which record the model at each epoch. These are called in the same way as above:

F_trajectory, Σ_mean, acceptance_rate, loss = train_rwm(F0, data, Σ0, opts);

The functions are analogously named:

ARFF.train_rwm — Function

train_rwm(F₀::FourierModel{TB,TR,TW}, data::DataSet{TB,TR,TW}, Σ::TM, options::ARFFOptions; show_progress=true, record_loss=true) where {TB<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TM<:AbstractMatrix}

Train the Fourier feature model using a random walk Metropolis exploration strategy. Returns the entire trajectory of models during training.

Fields

F₀ - The initial state of the FourierModel to be trained

data- The DataSet training data

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

train_rwm(F₀::FourierModel{TB,TR,TW}, batched_data::Vector{DataSet{TB,TR,TW}}, Σ::TM, options::ARFFOptions; show_progress=true) where {TB<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TM<:AbstractMatrix}

Train the Fourier feature model using a random walk Metropolis exploration strategy with batched data, which is cycled through from epoch to epoch. Returns the entire trajectory of models during training.

Fields

F₀ - The initial state of the FourierModel to be trained

batched_data- A vector of DataSet training data sets, for the purpose of minibatching. These are presumed to all be the same size.

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record

train_rwm(F₀::FourierModel{TB,TR,TW}, data::DataSet{TB,TR,TW}, batch_size::TI, Σ::TM, options::ARFFOptions; show_progress=true, record_loss=true) where {TB<:Complex,TR<:AbstractFloat,TW<:AbstractArray{TR},TM<:AbstractMatrix,TI<:Integer}

Train the Fourier feature model using a random walk Metropolis exploration strategy with minibatching, randomly subsampling at each epoch. Returns the entire trajectory of models during training.

F₀ - The initial state of the FourierModel to be trained

batch_size- Minibatch size

Σ - Initial covariance matrix for RWM proposals

options - ARFFOptions structure specifcying the number epochs, proposal step size, etc.

show_progress=true - Display training progress using ProgressMeter

record_loss=true - Evaluate the specified loss function at each epoch and record